I have the ActivityTrackerMain running.  you can view the user's goal as well as store a new one.
Currently I am working on viewing a user's goal... I can only get the
first line because it keeps reading the switch to the next line weird...
I'll continue to work on it.

Also I need to work on refactoring, because right now my code is pretty
nasty looking.

Feel free to test out the Main

+++ seis635team6/src/activity_tracker/TextType.java
@@ -4,8 +4,8 @@ public abstract class TextType {
 	public String getUsername(){
 		return null;
 	}
-	public Long getStartdate(){
-		return null;
+	public int getStartdate(){
+		return (Integer) null;
 	}
 
 }
